+++ CronJob/CronJob.cs
@@ -9,7 +9,7 @@ public class CronJob : BaseUnityPlugin
 {
   public const string GUID = "cron_job";
   public const string NAME = "Cron Job";
-  public const string VERSION = "1.5";
+  public const string VERSION = "1.6";
   private static ManualLogSource? Logs;
   public static ManualLogSource Log => Logs!;
   public void Awake()